refactor: move node to separate module
