+++ ethers-solc/src/resolver/mod.rs
@@ -46,16 +46,18 @@
 //! [version pragma](https://docs.soliditylang.org/en/develop/layout-of-source-files.html#version-pragma),
 //! which is defined on a per source file basis.
 
-use crate::{error::Result, utils, IncludePaths, ProjectPathsConfig, SolcError, Source, Sources};
-use parse::{SolData, SolDataUnit, SolImport};
+use crate::{error::Result, IncludePaths, ProjectPathsConfig, SolcError, Source, Sources};
+use node::{DisplayNode, Node};
+use parse::SolData;
 use rayon::prelude::*;
 use semver::VersionReq;
 use std::{
     collections::{HashMap, HashSet, VecDeque},
-    fmt, io,
+    io,
     path::{Path, PathBuf},
 };
 
+mod node;
 mod parse;
 mod tree;
 
@@ -816,73 +818,6 @@ impl VersionedSources {
             sources_by_version.insert(solc, (version, sources));
         }
         Ok(sources_by_version)
-    }
-}
-
-#[derive(Debug)]
-pub struct Node {
-    /// path of the solidity  file
-    path: PathBuf,
-    /// content of the solidity file
-    source: Source,
-    /// parsed data
-    data: SolData,
-}
-
-impl Node {
-    /// Reads the content of the file and returns a [Node] containing relevant information
-    pub fn read(file: impl AsRef<Path>) -> Result<Self> {
-        let file = file.as_ref();
-        let source = Source::read(file).map_err(|err| {
-            if !err.path().exists() && err.path().is_symlink() {
-                SolcError::ResolveBadSymlink(err)
-            } else {
-                SolcError::Resolve(err)
-            }
-        })?;
-        let data = SolData::parse(source.as_ref(), file);
-        Ok(Self { path: file.to_path_buf(), source, data })
-    }
-
-    pub fn content(&self) -> &str {
-        &self.source.content
-    }
-
-    pub fn imports(&self) -> &Vec<SolDataUnit<SolImport>> {
-        &self.data.imports
-    }
-
-    pub fn version(&self) -> &Option<SolDataUnit<String>> {
-        &self.data.version
-    }
-
-    pub fn experimental(&self) -> &Option<SolDataUnit<String>> {
-        &self.data.experimental
-    }
-
-    pub fn license(&self) -> &Option<SolDataUnit<String>> {
-        &self.data.license
-    }
-
-    pub fn unpack(&self) -> (&PathBuf, &Source) {
-        (&self.path, &self.source)
-    }
-}
-
-/// Helper type for formatting a node
-pub(crate) struct DisplayNode<'a> {
-    node: &'a Node,
-    root: &'a PathBuf,
-}
-
-impl<'a> fmt::Display for DisplayNode<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let path = utils::source_name(&self.node.path, self.root);
-        write!(f, "{}", path.display())?;
-        if let Some(ref v) = self.node.data.version {
-            write!(f, " {}", v.data())?;
-        }
-        Ok(())
     }
 }
